api/v2: remove admin routes, cleanup test files
