+++ api/v2/api_test.go
@@ -1,17 +1,14 @@
 package v2
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
 	"io"
 	"io/ioutil"
-	"mime/multipart"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
-	"os"
 	"strings"
 	"testing"
 	"time"
@@ -22,8 +19,6 @@ import (
 	"github.com/RTradeLtd/config"
 	"github.com/RTradeLtd/database"
 	"github.com/RTradeLtd/database/models"
-	pbOrch "github.com/RTradeLtd/grpc/ipfs-orchestrator"
-	pbLensResp "github.com/RTradeLtd/grpc/lens/response"
 	"github.com/RTradeLtd/rtfs"
 	"github.com/c2h5oh/datasize"
 	"github.com/gin-gonic/gin"
@@ -220,218 +215,6 @@ func Test_API_Setup(t *testing.T) {
 	}
 }
 
-func Test_API_Routes_Lens(t *testing.T) {
-	// load configuration
-	cfg, err := config.LoadConfig("../../testenv/config.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	db, err := loadDatabase(cfg)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// setup fake mock clients
-	fakeLens := &mocks.FakeIndexerAPIClient{}
-	fakeOrch := &mocks.FakeServiceClient{}
-	fakeSigner := &mocks.FakeSignerClient{}
-
-	api, _, err := setupAPI(fakeLens, fakeOrch, fakeSigner, cfg, db)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// test lens index - missing post form
-	// /api/v2/lens/index
-	var apiResp apiResponse
-	urlValues := url.Values{}
-	urlValues.Add("object_type", "ipld")
-	if err := sendRequest(
-		api, "POST", "/api/v2/lens/index", 400, nil, urlValues, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-
-	// test lens index - valid object type
-	// /api/v2/lens/index
-	var mapAPIResp mapAPIResponse
-	urlValues = url.Values{}
-	urlValues.Add("object_type", "ipld")
-	urlValues.Add("object_identifier", hash)
-	// setup our mock index response
-	fakeLens.IndexReturnsOnCall(0, &pbLensResp.Index{
-		Id:       "fakeid",
-		Keywords: []string{"protocols", "minivan"},
-	}, nil)
-	if err := sendRequest(
-		api, "POST", "/api/v2/lens/index", 200, nil, urlValues, &mapAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	if mapAPIResp.Code != 200 {
-		t.Fatal("bad api response status code from /api/v2/lens/index")
-	}
-
-	// test lens index - invalid object type
-	// /api/v2/lens/index
-	urlValues = url.Values{}
-	urlValues.Add("object_type", "storj")
-	urlValues.Add("object_identifier", hash)
-	if err := sendRequest(
-		api, "POST", "/api/v2/lens/index", 400, nil, urlValues, nil,
-	); err != nil {
-		t.Fatal(err)
-	}
-
-	// test lens index - bad format hash
-	// /api/v2/lens/index
-	urlValues = url.Values{}
-	urlValues.Add("object_type", "ipld")
-	urlValues.Add("object_identifier", "notarealipfshash")
-	if err := sendRequest(
-		api, "POST", "/api/v2/lens/index", 400, nil, urlValues, nil,
-	); err != nil {
-		t.Fatal(err)
-	}
-
-	// test lens search - with no objects
-	// /api/v2/lens/search
-	apiResp = apiResponse{}
-	fakeLens.SearchReturnsOnCall(0, nil, nil)
-	urlValues = url.Values{}
-	urlValues.Add("keywords", "notarealsearch")
-	if err := sendRequest(
-		api, "POST", "/api/v2/lens/search", 400, nil, urlValues, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	if apiResp.Code != 400 {
-		t.Fatal("bad api response code from /api/v2/lens/search")
-	}
-	if apiResp.Response != "no results found" {
-		t.Fatal("failed to correctly detect no results found")
-	}
-
-	// test lens search - with objects
-	// /api/v2/lens/search
-	// setup our mock search response
-	var lensSearchAPIResp lensSearchAPIResponse
-	obj := pbLensResp.Object{
-		Name:     hash,
-		MimeType: "application/pdf",
-		Category: "documents",
-	}
-	var objs []*pbLensResp.Object
-	objs = append(objs, &obj)
-	fakeLens.SearchReturnsOnCall(1, &pbLensResp.Results{
-		Objects: objs,
-	}, nil)
-	urlValues = url.Values{}
-	urlValues.Add("keywords", "minivan")
-	urlValues.Add("keywords", "protocols")
-	if err := sendRequest(
-		api, "POST", "/api/v2/lens/search", 200, nil, urlValues, &lensSearchAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	fmt.Println(lensSearchAPIResp)
-	if lensSearchAPIResp.Code != 200 {
-		t.Fatal("bad api response code from /api/v2/lens/search")
-	}
-	if lensSearchAPIResp.Response[0]["name"] != hash {
-		t.Fatal("failed to search for correct hash")
-	}
-	if lensSearchAPIResp.Response[0]["category"] != "documents" {
-		t.Fatal("failed to search for correct category")
-	}
-	if lensSearchAPIResp.Response[0]["mimeType"] != "application/pdf" {
-		t.Fatal("failed to search for correct mimetpye")
-	}
-}
-
-func Test_API_Routes_Payments(t *testing.T) {
-	// load configuration
-	cfg, err := config.LoadConfig("../../testenv/config.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	db, err := loadDatabase(cfg)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// setup fake mock clients
-	fakeLens := &mocks.FakeIndexerAPIClient{}
-	fakeOrch := &mocks.FakeServiceClient{}
-	fakeSigner := &mocks.FakeSignerClient{}
-
-	api, testRecorder, err := setupAPI(fakeLens, fakeOrch, fakeSigner, cfg, db)
-	if err != nil {
-		t.Fatal(err)
-	}
-	// test basic dash payment
-	// /api/v2/payments/create/dash
-	testRecorder = httptest.NewRecorder()
-	req := httptest.NewRequest("POST", "/api/v2/payments/create/dash", nil)
-	req.Header.Add("Authorization", authHeader)
-	urlValues := url.Values{}
-	urlValues.Add("credit_value", "10")
-	req.PostForm = urlValues
-	api.r.ServeHTTP(testRecorder, req)
-
-	// test request signed payment message - rtc
-	// /api/v2/payments/request
-	testRecorder = httptest.NewRecorder()
-	req = httptest.NewRequest("POST", "/api/v2/payments/request", nil)
-	req.Header.Add("Authorization", authHeader)
-	urlValues = url.Values{}
-	urlValues.Add("payment_type", "rtc")
-	urlValues.Add("sender_address", "0x0")
-	urlValues.Add("credit_value", "10")
-	req.PostForm = urlValues
-	api.r.ServeHTTP(testRecorder, req)
-
-	// test request signed payment message - eth
-	// /api/v2/payments/request
-	testRecorder = httptest.NewRecorder()
-	req = httptest.NewRequest("POST", "/api/v2/payments/request", nil)
-	req.Header.Add("Authorization", authHeader)
-	urlValues = url.Values{}
-	urlValues.Add("payment_type", "eth")
-	urlValues.Add("sender_address", "0x0")
-	urlValues.Add("credit_value", "10")
-	req.PostForm = urlValues
-	api.r.ServeHTTP(testRecorder, req)
-
-	// test payment confirmation
-	// /api/v2/payments/confirm
-	testRecorder = httptest.NewRecorder()
-	req = httptest.NewRequest("POST", "/api/v2/payments/confirm", nil)
-	req.Header.Add("Authorization", authHeader)
-	urlValues = url.Values{}
-	urlValues.Add("payment_number", "10")
-	urlValues.Add("tx_hash", "0x1")
-	req.PostForm = urlValues
-	api.r.ServeHTTP(testRecorder, req)
-
-	// test valid deposit address
-	args := []string{"eth", "rtc", "btc", "ltc", "xmr", "dash"}
-	for _, v := range args {
-		if err := sendRequest(
-			api, "GET", "/api/v2/payments/deposit/address/"+v, 200, nil, nil, nil,
-		); err != nil {
-			t.Fatal(err)
-		}
-	}
-
-	// test invalid deposit address
-	if err := sendRequest(
-		api, "GET", "/api/v2/payments/deposit/address/invalidType", 400, nil, nil, nil,
-	); err != nil {
-		t.Fatal(err)
-	}
-}
-
 func Test_API_Routes_Misc(t *testing.T) {
 	// load configuration
 	cfg, err := config.LoadConfig("../../testenv/config.json")
@@ -477,1255 +260,8 @@ func Test_API_Routes_Misc(t *testing.T) {
 	); err != nil {
 		t.Fatal(err)
 	}
-
-	// test mini create bucket
-	// /api/v2/admin/mini/create/bucket
-	if err := os.Setenv("MINI_SSL_ENABLE", "false"); err != nil {
-		t.Fatal(err)
-	}
-	urlValues := url.Values{}
-	urlValues.Add("bucket_name", "filesuploadbucket2")
-	if err := sendRequest(
-		api, "POST", "/api/v2/admin/mini/create/bucket", 200, nil, urlValues, nil,
-	); err != nil {
-		t.Fatal(err)
-	}
 }
 
-func Test_API_Routes_IPFS_Public(t *testing.T) {
-	// load configuration
-	cfg, err := config.LoadConfig("../../testenv/config.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	db, err := loadDatabase(cfg)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// setup fake mock clients
-	fakeLens := &mocks.FakeIndexerAPIClient{}
-	fakeOrch := &mocks.FakeServiceClient{}
-	fakeSigner := &mocks.FakeSignerClient{}
-
-	api, testRecorder, err := setupAPI(fakeLens, fakeOrch, fakeSigner, cfg, db)
-	if err != nil {
-		t.Fatal(err)
-	}
-	// add a file normally
-	// /api/v2/ipfs/public/file/add
-	bodyBuf := &bytes.Buffer{}
-	bodyWriter := multipart.NewWriter(bodyBuf)
-	fileWriter, err := bodyWriter.CreateFormFile("file", "../../testenv/config.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	fh, err := os.Open("../../testenv/config.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer fh.Close()
-	if _, err = io.Copy(fileWriter, fh); err != nil {
-		t.Fatal(err)
-	}
-	bodyWriter.Close()
-	testRecorder = httptest.NewRecorder()
-	req := httptest.NewRequest("POST", "/api/v2/ipfs/public/file/add", bodyBuf)
-	req.Header.Add("Authorization", authHeader)
-	req.Header.Add("Content-Type", bodyWriter.FormDataContentType())
-	urlValues := url.Values{}
-	urlValues.Add("hold_time", "5")
-	req.PostForm = urlValues
-	api.r.ServeHTTP(testRecorder, req)
-	if testRecorder.Code != 200 {
-		t.Fatal("bad http status code recovered from /api/v2/ipfs/public/file/add")
-	}
-	var apiResp apiResponse
-	// unmarshal the response
-	bodyBytes, err := ioutil.ReadAll(testRecorder.Result().Body)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err = json.Unmarshal(bodyBytes, &apiResp); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if apiResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/ipfs/public/file/add")
-	}
-	hash = apiResp.Response
-
-	// add a file advanced
-	// /api/v2/ipfs/public/file/add/advanced
-	bodyBuf = &bytes.Buffer{}
-	bodyWriter = multipart.NewWriter(bodyBuf)
-	fileWriter, err = bodyWriter.CreateFormFile("file", "../../testenv/config.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	fh, err = os.Open("../../testenv/config.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer fh.Close()
-	if _, err = io.Copy(fileWriter, fh); err != nil {
-		t.Fatal(err)
-	}
-	bodyWriter.Close()
-	testRecorder = httptest.NewRecorder()
-	req = httptest.NewRequest("POST", "/api/v2/ipfs/public/file/add/advanced", bodyBuf)
-	req.Header.Add("Authorization", authHeader)
-	req.Header.Add("Content-Type", bodyWriter.FormDataContentType())
-	urlValues = url.Values{}
-	urlValues.Add("hold_time", "5")
-	urlValues.Add("passphrase", "password123")
-	req.PostForm = urlValues
-	api.r.ServeHTTP(testRecorder, req)
-	if testRecorder.Code != 200 {
-		t.Fatal("bad http status code recovered from /api/v2/ipfs/public/file/add/advanced")
-	}
-	apiResp = apiResponse{}
-	// unmarshal the response
-	bodyBytes, err = ioutil.ReadAll(testRecorder.Result().Body)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err = json.Unmarshal(bodyBytes, &apiResp); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if apiResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/ipfs/public/file/add/advanced")
-	}
-
-	// test pinning
-	// /api/v2/ipfs/public/pin
-	apiResp = apiResponse{}
-	urlValues = url.Values{}
-	urlValues.Add("hold_time", "5")
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipfs/public/pin/"+hash, 200, nil, urlValues, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if apiResp.Code != 200 {
-		t.Fatal("bad api status code from  /api/v2/ipfs/public/pin")
-	}
-
-	// test pin check
-	// /api/v2/ipfs/public/check/pin
-	var boolAPIResp boolAPIResponse
-	if err := sendRequest(
-		api, "GET", "/api/v2/ipfs/public/pin/check/"+hash, 200, nil, nil, &boolAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if boolAPIResp.Code != 200 {
-		t.Fatal("bad api status code from  /api/v2/ipfs/public/check/pin")
-	}
-
-	// test pubsub publish
-	// /api/v2/ipfs/pubsub/publish/topic
-	urlValues = url.Values{}
-	urlValues.Add("message", "bar")
-	var mapAPIResp mapAPIResponse
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipfs/public/pubsub/publish/foo", 200, nil, urlValues, &mapAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if mapAPIResp.Code != 200 {
-		t.Fatal("bad api status code from  /api/v2/pubsub/publish/topic")
-	}
-	if mapAPIResp.Response["topic"] != "foo" {
-		t.Fatal("bad response")
-	}
-	if mapAPIResp.Response["message"] != "bar" {
-		t.Fatal("bad response")
-	}
-
-	// test object stat
-	// /api/v2/ipfs/stat
-	var interfaceAPIResp interfaceAPIResponse
-	if err := sendRequest(
-		api, "GET", "/api/v2/ipfs/public/stat/"+hash, 200, nil, nil, &interfaceAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	if interfaceAPIResp.Code != 200 {
-		t.Fatal("bad response status code from /api/v2/ipfs/public/stat")
-	}
-
-	// test get dag
-	// /api/v2/ipfs/public/dag
-	interfaceAPIResp = interfaceAPIResponse{}
-	if err := sendRequest(
-		api, "GET", "/api/v2/ipfs/public/dag/"+hash, 200, nil, nil, &interfaceAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	if interfaceAPIResp.Code != 200 {
-		t.Fatal("bad response status code from /api/v2/ipfs/public/dag/")
-	}
-
-	// test download
-	// /api/v2/ipfs/utils/download
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipfs/utils/download/"+hash, 200, nil, nil, nil,
-	); err != nil {
-		t.Fatal(err)
-	}
-
-	// test public network beam
-	// /api/v2/ipfs/utils/laser/beam
-	urlValues = url.Values{}
-	urlValues.Add("source_network", "public")
-	urlValues.Add("destination_network", "public")
-	urlValues.Add("content_hash", hash)
-	urlValues.Add("passphrase", "password123")
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipfs/utils/laser/beam", 200, nil, urlValues, nil,
-	); err != nil {
-		t.Fatal(err)
-	}
-}
-
-func Test_API_Routes_IPFS_Private(t *testing.T) {
-	// load configuration
-	cfg, err := config.LoadConfig("../../testenv/config.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	db, err := loadDatabase(cfg)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// setup fake mock clients
-	fakeLens := &mocks.FakeIndexerAPIClient{}
-	fakeOrch := &mocks.FakeServiceClient{}
-	fakeSigner := &mocks.FakeSignerClient{}
-
-	api, testRecorder, err := setupAPI(fakeLens, fakeOrch, fakeSigner, cfg, db)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	nm := models.NewHostedIPFSNetworkManager(db)
-
-	// create private network - failure missing name
-	// /api/v2/ipfs/private/new
-	var apiResp apiResponse
-	urlValues := url.Values{}
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipfs/private/network/new", 400, nil, nil, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	if apiResp.Code != 400 {
-		t.Fatal("bad api status code from /api/v2/ipfs/private/network/new")
-	}
-	if apiResp.Response != "network_name not present" {
-		t.Fatal("failed to detect missing network_name field")
-	}
-
-	// create private network - failure name is PUBLIC
-	// /api/v2/ipfs/private/new
-	apiResp = apiResponse{}
-	urlValues = url.Values{}
-	urlValues.Add("network_name", "PUBLIC")
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipfs/private/network/new", 400, nil, nil, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	if apiResp.Code != 400 {
-		t.Fatal("bad api status code from /api/v2/ipfs/private/network/new")
-	}
-
-	// create private network - failure name is public
-	// /api/v2/ipfs/private/new
-	apiResp = apiResponse{}
-	urlValues = url.Values{}
-	urlValues.Add("network_name", "public")
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipfs/private/network/new", 400, nil, nil, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	if apiResp.Code != 400 {
-		t.Fatal("bad api status code from /api/v2/ipfs/private/network/new")
-	}
-
-	// create private network
-	// /api/v2/ipfs/private/new
-	var mapAPIResp mapAPIResponse
-	urlValues = url.Values{}
-	urlValues.Add("network_name", "abc123")
-	fakeOrch.StartNetworkReturnsOnCall(0, &pbOrch.StartNetworkResponse{Api: "/ip4/127.0.0.1/tcp/5001", SwarmKey: testSwarmKey}, nil)
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipfs/private/network/new", 200, nil, urlValues, &mapAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	if mapAPIResp.Code != 200 {
-		t.Fatal("bad api response status code from /api/v2/ipfs/private/new")
-	}
-	if mapAPIResp.Response["network_name"] != "abc123" {
-		t.Fatal("failed to retrieve correct network name")
-	}
-	if mapAPIResp.Response["api_url"] != "/ip4/127.0.0.1/tcp/5001" {
-		t.Fatal("failed to retrieve correct api url")
-	}
-	if mapAPIResp.Response["swarm_key"] != testSwarmKey {
-		t.Fatal("failed to get correct swarm key")
-	}
-	if err := nm.UpdateNetworkByName("abc123", map[string]interface{}{
-		"api_url": cfg.IPFS.APIConnection.Host + ":" + cfg.IPFS.APIConnection.Port,
-	}); err != nil {
-		t.Fatal(err)
-	}
-
-	// create private network with parameters
-	// /api/v2/ipfs/private/network
-	mapAPIResp = mapAPIResponse{}
-	urlValues = url.Values{}
-	urlValues.Add("network_name", "xyz123")
-	urlValues.Add("swarm_key", testSwarmKey)
-	urlValues.Add("bootstrap_peers", testBootstrapPeer1)
-	urlValues.Add("bootstrap_peers", testBootstrapPeer2)
-	urlValues.Add("users", "testuser")
-	urlValues.Add("users", "testuser2")
-	fakeOrch.StartNetworkReturnsOnCall(1, &pbOrch.StartNetworkResponse{Api: "/ip4/127.0.0.1/tcp/5002", SwarmKey: "swarmStorm"}, nil)
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipfs/private/network/new", 200, nil, urlValues, &mapAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	if mapAPIResp.Code != 200 {
-		t.Fatal("bad api response status code from /api/v2/ipfs/private/new")
-	}
-	if mapAPIResp.Response["network_name"] != "xyz123" {
-		t.Fatal("failed to retrieve correct network name")
-	}
-	if mapAPIResp.Response["api_url"] != "/ip4/127.0.0.1/tcp/5002" {
-		t.Fatal("failed to retrieve correct api url")
-	}
-	if mapAPIResp.Response["swarm_key"] != "swarmStorm" {
-		t.Fatal("failed to get correct swarm key")
-	}
-
-	// get private network information
-	// /api/v2/ipfs/private/network/:name
-	var interfaceAPIResp interfaceAPIResponse
-	if err := sendRequest(
-		api, "GET", "/api/v2/ipfs/private/network/abc123", 200, nil, nil, &interfaceAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	if interfaceAPIResp.Code != 200 {
-		t.Fatal("bad api response status code from /api/v2/ipfs/private/network/abc123")
-	}
-
-	// get all authorized private networks
-	// /api/v2/ipfs/private/networks
-	var stringSliceAPIResp stringSliceAPIResponse
-	if err := sendRequest(
-		api, "GET", "/api/v2/ipfs/private/networks", 200, nil, nil, &stringSliceAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	if stringSliceAPIResp.Code != 200 {
-		t.Fatal("bad api response status code from /api/v2/ipfs/private/networks")
-	}
-	if len(stringSliceAPIResp.Response) == 0 {
-		t.Fatal("failed to find any from /api/v2/ipfs/private/networks")
-	}
-	var found bool
-	for _, v := range stringSliceAPIResp.Response {
-		if v == "abc123" {
-			found = true
-			break
-		}
-	}
-	if !found {
-		t.Fatal("failed to find correct network from /api/v2/ipfs/private/networks")
-	}
-
-	// stop private network
-	// /api/v2/ipfs/private/network/stop
-	// for now until we implement proper grpc testing, this will fail
-	mapAPIResp = mapAPIResponse{}
-	urlValues = url.Values{}
-	urlValues.Add("network_name", "abc123")
-	fakeOrch.StopNetworkReturnsOnCall(0, &pbOrch.Empty{}, nil)
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipfs/private/network/stop", 200, nil, urlValues, &mapAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	if mapAPIResp.Code != 200 {
-		t.Fatal("bad api resposne code from /api/v2/ipfs/private/network/stop")
-	}
-	if mapAPIResp.Response["state"] != "stopped" {
-		t.Fatal("failed to stop network")
-	}
-
-	// start private network
-	// /api/v2/ipfs/private/network/start
-	// for now until we implement proper grpc testing, this will fail
-	mapAPIResp = mapAPIResponse{}
-	urlValues = url.Values{}
-	urlValues.Add("network_name", "abc123")
-	fakeOrch.StartNetworkReturnsOnCall(2, &pbOrch.StartNetworkResponse{Api: "test", SwarmKey: "test"}, nil)
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipfs/private/network/start", 200, nil, urlValues, &mapAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	if mapAPIResp.Code != 200 {
-		t.Fatal("bad api resposne code from /api/v2/ipfs/private/network/stop")
-	}
-	if mapAPIResp.Response["state"] != "started" {
-		t.Fatal("failed to stop network")
-	}
-
-	// add a file normally
-	// /api/v2/ipfs/private/file/add
-	bodyBuf := &bytes.Buffer{}
-	bodyWriter := multipart.NewWriter(bodyBuf)
-	fileWriter, err := bodyWriter.CreateFormFile("file", "../../testenv/config.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	fh, err := os.Open("../../testenv/config.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer fh.Close()
-	if _, err = io.Copy(fileWriter, fh); err != nil {
-		t.Fatal(err)
-	}
-	bodyWriter.Close()
-	testRecorder = httptest.NewRecorder()
-	req := httptest.NewRequest("POST", "/api/v2/ipfs/private/file/add", bodyBuf)
-	req.Header.Add("Authorization", authHeader)
-	req.Header.Add("Content-Type", bodyWriter.FormDataContentType())
-	urlValues = url.Values{}
-	urlValues.Add("hold_time", "5")
-	urlValues.Add("network_name", "abc123")
-	req.PostForm = urlValues
-	api.r.ServeHTTP(testRecorder, req)
-	if testRecorder.Code != 200 {
-		t.Fatal("bad http status code recovered from /api/v2/ipfs/private/file/add")
-	}
-	apiResp = apiResponse{}
-	// unmarshal the response
-	bodyBytes, err := ioutil.ReadAll(testRecorder.Result().Body)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err = json.Unmarshal(bodyBytes, &apiResp); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if apiResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/ipfs/private/file/add")
-	}
-	hash = apiResp.Response
-
-	// add a file advanced
-	// /api/v2/ipfs/private/file/add/advanced
-	bodyBuf = &bytes.Buffer{}
-	bodyWriter = multipart.NewWriter(bodyBuf)
-	fileWriter, err = bodyWriter.CreateFormFile("file", "../../testenv/config.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	fh, err = os.Open("../../testenv/config.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer fh.Close()
-	if _, err = io.Copy(fileWriter, fh); err != nil {
-		t.Fatal(err)
-	}
-	bodyWriter.Close()
-	testRecorder = httptest.NewRecorder()
-	req = httptest.NewRequest("POST", "/api/v2/ipfs/private/file/add/advanced", bodyBuf)
-	req.Header.Add("Authorization", authHeader)
-	req.Header.Add("Content-Type", bodyWriter.FormDataContentType())
-	urlValues = url.Values{}
-	urlValues.Add("hold_time", "5")
-	urlValues.Add("passphrase", "password123")
-	urlValues.Add("network_name", "abc123")
-	req.PostForm = urlValues
-	api.r.ServeHTTP(testRecorder, req)
-	if testRecorder.Code != 200 {
-		t.Fatal("bad http status code recovered from /api/v2/ipfs/private/file/add/advanced")
-	}
-	apiResp = apiResponse{}
-	// unmarshal the response
-	bodyBytes, err = ioutil.ReadAll(testRecorder.Result().Body)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err = json.Unmarshal(bodyBytes, &apiResp); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if apiResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/ipfs/private/file/add/advanced")
-	}
-
-	// test pinning
-	// /api/v2/ipfs/private/pin
-	apiResp = apiResponse{}
-	urlValues = url.Values{}
-	urlValues.Add("hold_time", "5")
-	urlValues.Add("network_name", "abc123")
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipfs/private/pin/"+hash, 200, nil, urlValues, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if apiResp.Code != 200 {
-		t.Fatal("bad api status code from  /api/v2/ipfs/private/pin")
-	}
-
-	// test pin check
-	// /api/v2/ipfs/private/check/pin
-	var boolAPIResp boolAPIResponse
-	if err := sendRequest(
-		api, "GET", "/api/v2/ipfs/private/pin/check/"+hash+"/abc123", 200, nil, nil, &boolAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if boolAPIResp.Code != 200 {
-		t.Fatal("bad api status code from  /api/v2/ipfs/private/check/pin")
-	}
-
-	// test pubsub publish
-	// /api/v2/ipfs/private/publish/topic
-	mapAPIResp = mapAPIResponse{}
-	urlValues = url.Values{}
-	urlValues.Add("message", "bar")
-	urlValues.Add("network_name", "abc123")
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipfs/private/pubsub/publish/foo", 200, nil, urlValues, &mapAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if mapAPIResp.Code != 200 {
-		t.Fatal("bad api status code from  /api/v2/ipfs/private/pubsub/publish/topic")
-	}
-	if mapAPIResp.Response["topic"] != "foo" {
-		t.Fatal("bad response")
-	}
-	if mapAPIResp.Response["message"] != "bar" {
-		t.Fatal("bad response")
-	}
-
-	// test object stat
-	// /api/v2/ipfs/private/stat
-	interfaceAPIResp = interfaceAPIResponse{}
-	if err := sendRequest(
-		api, "GET", "/api/v2/ipfs/private/stat/"+hash+"/abc123", 200, nil, nil, &interfaceAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	if interfaceAPIResp.Code != 200 {
-		t.Fatal("bad response status code from /api/v2/ipfs/private/stat")
-	}
-
-	// test get dag
-	// /api/v2/ipfs/private/dag
-	interfaceAPIResp = interfaceAPIResponse{}
-	if err := sendRequest(
-		api, "GET", "/api/v2/ipfs/private/dag/"+hash+"/abc123", 200, nil, nil, &interfaceAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	if interfaceAPIResp.Code != 200 {
-		t.Fatal("bad response status code from /api/v2/ipfs/private/dag/")
-	}
-
-	// test download
-	// /api/v2/ipfs/utils/download
-	urlValues = url.Values{}
-	urlValues.Add("network_name", "abc123")
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipfs/utils/download/"+hash, 200, nil, urlValues, nil,
-	); err != nil {
-		t.Fatal(err)
-	}
-
-	// test get authorized networks
-	// /api/v2/ipfs/private/networks
-	interfaceAPIResp = interfaceAPIResponse{}
-	if err := sendRequest(
-		api, "GET", "/api/v2/ipfs/private/networks", 200, nil, nil, &interfaceAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	if interfaceAPIResp.Code != 200 {
-		t.Fatal("bad response status code from /api/v2/ipfs/private/networks/")
-	}
-
-	// test get authorized networks
-	// /api/v2/ipfs/private/networks
-	interfaceAPIResp = interfaceAPIResponse{}
-	if err := sendRequest(
-		api, "GET", "/api/v2/ipfs/private/uploads/abc123", 200, nil, nil, &interfaceAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	if interfaceAPIResp.Code != 200 {
-		t.Fatal("bad response status code from /api/v2/ipfs/private/uploads")
-	}
-
-	// test private network beam - source private, dest public
-	urlValues = url.Values{}
-	urlValues.Add("source_network", "abc123")
-	urlValues.Add("destination_network", "public")
-	urlValues.Add("content_hash", hash)
-	urlValues.Add("passphrase", "password123")
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipfs/utils/laser/beam", 200, nil, urlValues, nil,
-	); err != nil {
-		t.Fatal(err)
-	}
-
-	// test private network beam - source public, dest private
-	// /api/v2/ipfs/utils/laser/beam
-	urlValues = url.Values{}
-	urlValues.Add("source_network", "public")
-	urlValues.Add("destination_network", "abc123")
-	urlValues.Add("content_hash", hash)
-	urlValues.Add("passphrase", "password123")
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipfs/utils/laser/beam", 200, nil, urlValues, nil,
-	); err != nil {
-		t.Fatal(err)
-	}
-
-	// test private network beam - source private, dest private
-	// /api/v2/ipfs/utils/laser/beam
-	urlValues = url.Values{}
-	urlValues.Add("source_network", "abc123")
-	urlValues.Add("destination_network", "abc123")
-	urlValues.Add("content_hash", hash)
-	urlValues.Add("passphrase", "password123")
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipfs/utils/laser/beam", 200, nil, urlValues, nil,
-	); err != nil {
-		t.Fatal(err)
-	}
-
-	// remove private network
-	// /api/v2/ipfs/private/network/remove
-	// for now until we implement proper grpc testing, this will fail
-	mapAPIResp = mapAPIResponse{}
-	urlValues = url.Values{}
-	urlValues.Add("network_name", "abc123")
-	fakeOrch.RemoveNetworkReturnsOnCall(0, &pbOrch.Empty{}, nil)
-	if err := sendRequest(
-		api, "DELETE", "/api/v2/ipfs/private/network/remove", 200, nil, urlValues, &mapAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	if mapAPIResp.Code != 200 {
-		t.Fatal("bad api response status code from /api/v2/ipfs/private/network/remove")
-	}
-	if mapAPIResp.Response["state"] != "removed" {
-		t.Fatal("failed to remove network")
-	}
-}
-
-func Test_API_Routes_IPNS(t *testing.T) {
-	// load configuration
-	cfg, err := config.LoadConfig("../../testenv/config.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	db, err := loadDatabase(cfg)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// setup fake mock clients
-	fakeLens := &mocks.FakeIndexerAPIClient{}
-	fakeOrch := &mocks.FakeServiceClient{}
-	fakeSigner := &mocks.FakeSignerClient{}
-
-	api, testRecorder, err := setupAPI(fakeLens, fakeOrch, fakeSigner, cfg, db)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	um := models.NewUserManager(db)
-	um.AddIPFSKeyForUser("testuser", "mytestkey", "suchkeymuchwow")
-
-	// test get ipns records
-	// /api/v2/ipns/records
-	testRecorder = httptest.NewRecorder()
-	req := httptest.NewRequest("GET", "/api/v2/ipns/records", nil)
-	req.Header.Add("Authorization", authHeader)
-	api.r.ServeHTTP(testRecorder, req)
-	if testRecorder.Code != 200 {
-		t.Fatal("bad http status code from /api/v2/ipns/records")
-	}
-
-	// test ipns publishing (public) - bad hash
-	// /api/v2/ipns/public/publish/details
-	var apiResp apiResponse
-	urlValues := url.Values{}
-	urlValues.Add("hash", "notavalidipfshash")
-	urlValues.Add("life_time", "24h")
-	urlValues.Add("ttl", "1h")
-	urlValues.Add("key", "mytestkey")
-	urlValues.Add("resolve", "true")
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipns/public/publish/details", 400, nil, urlValues, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-
-	// test ipns publishing (private) - bad hash
-	// /api/v2/ipns/private/publish/details
-	apiResp = apiResponse{}
-	urlValues = url.Values{}
-	urlValues.Add("hash", "notavalidipfshash")
-	urlValues.Add("life_time", "24h")
-	urlValues.Add("ttl", "1h")
-	urlValues.Add("key", "mytestkey")
-	urlValues.Add("resolve", "true")
-	urlValues.Add("network_name", "testnetwork")
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipns/private/publish/details", 400, nil, urlValues, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-
-	// test ipns publishing (public)
-	// api/v2/ipns/public/publish/details
-	apiResp = apiResponse{}
-	urlValues = url.Values{}
-	urlValues.Add("hash", hash)
-	urlValues.Add("life_time", "24h")
-	urlValues.Add("ttl", "1h")
-	urlValues.Add("key", "mytestkey")
-	urlValues.Add("resolve", "true")
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipns/public/publish/details", 200, nil, urlValues, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if apiResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/ipns/public/publish/details")
-	}
-
-	// test ipns publishing (private)
-	// api/v2/ipns/private/publish/details
-	// create a fake private network
-	apiResp = apiResponse{}
-	nm := models.NewHostedIPFSNetworkManager(db)
-	if _, err := nm.CreateHostedPrivateNetwork("testnetwork", "fakeswarmkey", nil, []string{"testuser"}); err != nil {
-		t.Fatal(err)
-	}
-	if err := um.AddIPFSNetworkForUser("testuser", "testnetwork"); err != nil {
-		t.Fatal(err)
-	}
-	urlValues = url.Values{}
-	urlValues.Add("hash", hash)
-	urlValues.Add("life_time", "24h")
-	urlValues.Add("ttl", "1h")
-	urlValues.Add("key", "mytestkey")
-	urlValues.Add("resolve", "true")
-	urlValues.Add("network_name", "testnetwork")
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipns/private/publish/details", 200, nil, urlValues, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if apiResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/ipns/private/publish/details")
-	}
-
-	// test get ipns records
-	// /api/v2/ipns/records
-	// spoof a fake record as we arent using the queues in this test
-	var ipnsAPIResp ipnsAPIResponse
-	im := models.NewIPNSManager(db)
-	if _, err := im.CreateEntry("fakekey", "fakehash", "fakekeyname", "public", "testuser", time.Minute, time.Minute); err != nil {
-		t.Fatal(err)
-	}
-	if err := sendRequest(
-		api, "GET", "/api/v2/ipns/records", 200, nil, nil, &ipnsAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if ipnsAPIResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/ipns/private/publish/details")
-	}
-	if len(*ipnsAPIResp.Response) == 0 {
-		t.Fatal("no records discovered")
-	}
-}
-
-func Test_API_Routes_Cluster(t *testing.T) {
-	// load configuration
-	cfg, err := config.LoadConfig("../../testenv/config.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	db, err := loadDatabase(cfg)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// setup fake mock clients
-	fakeLens := &mocks.FakeIndexerAPIClient{}
-	fakeOrch := &mocks.FakeServiceClient{}
-	fakeSigner := &mocks.FakeSignerClient{}
-
-	api, _, err := setupAPI(fakeLens, fakeOrch, fakeSigner, cfg, db)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// test cluster sync
-	// /api/v2/ipfs/cluster/sync/errors/local
-	var interfaceAPIResp interfaceAPIResponse
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipfs/cluster/sync/errors/local", 200, nil, nil, &interfaceAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if interfaceAPIResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/ipfs/cluster/sync/errors/local")
-	}
-
-	// test cluster pin
-	// /api/v2/ipfs/cluster/pin
-	var apiResp apiResponse
-	urlValues := url.Values{}
-	urlValues.Add("hold_time", "5")
-	if err := sendRequest(
-		api, "POST", "/api/v2/ipfs/cluster/pin/"+hash, 200, nil, urlValues, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if apiResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/ipfs/cluster/pin")
-	}
-	// manually pin since we aren't using queues
-	decoded, err := api.ipfsCluster.DecodeHashString(hash)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err = api.ipfsCluster.Pin(decoded); err != nil {
-		t.Fatal(err)
-	}
-
-	// test cluster local status
-	// /api/v2/ipfs/cluster/status/pin/local
-	interfaceAPIResp = interfaceAPIResponse{}
-	if err := sendRequest(
-		api, "GET", "/api/v2/ipfs/cluster/status/pin/local/"+hash, 200, nil, nil, &interfaceAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if interfaceAPIResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/ipfs/cluster/status/pin/local")
-	}
-
-	// test cluster local status
-	// /api/v2/ipfs/cluster/status/pin/global
-	interfaceAPIResp = interfaceAPIResponse{}
-	if err := sendRequest(
-		api, "GET", "/api/v2/ipfs/cluster/status/pin/global/"+hash, 200, nil, nil, &interfaceAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if interfaceAPIResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/ipfs/cluster/status/pin/global")
-	}
-
-	// test cluster local status
-	// /api/v2/ipfs/cluster/status/local
-	interfaceAPIResp = interfaceAPIResponse{}
-	if err := sendRequest(
-		api, "GET", "/api/v2/ipfs/cluster/status/local", 200, nil, nil, &interfaceAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if interfaceAPIResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/ipfs/cluster/status/local")
-	}
-}
-
-func Test_API_Routes_Frontend(t *testing.T) {
-	// load configuration
-	cfg, err := config.LoadConfig("../../testenv/config.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	db, err := loadDatabase(cfg)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// setup fake mock clients
-	fakeLens := &mocks.FakeIndexerAPIClient{}
-	fakeOrch := &mocks.FakeServiceClient{}
-	fakeSigner := &mocks.FakeSignerClient{}
-
-	api, testRecorder, err := setupAPI(fakeLens, fakeOrch, fakeSigner, cfg, db)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// test pin cost calculate
-	// /api/v2/frontend/cost/calculate/:hash/:holTime
-	var floatAPIResp floatAPIResponse
-	if err := sendRequest(
-		api, "GET", "/api/v2/frontend/cost/calculate/"+hash+"/5", 200, nil, nil, &floatAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	if floatAPIResp.Code != 200 {
-		t.Fatal("bad response code from /api/v2/frontend/cost/calculate")
-	}
-	if floatAPIResp.Response == 0 {
-		t.Fatal("failed to calculate cost /api/v2/frontend/cost/calculate")
-	}
-
-	// test file upload cost calculation
-	// /api/v2/frontend/cost/calculate/file
-	bodyBuf := &bytes.Buffer{}
-	bodyWriter := multipart.NewWriter(bodyBuf)
-	fileWriter, err := bodyWriter.CreateFormFile("file", "../../testenv/config.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	fh, err := os.Open("../../testenv/config.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer fh.Close()
-	if _, err = io.Copy(fileWriter, fh); err != nil {
-		t.Fatal(err)
-	}
-	bodyWriter.Close()
-	testRecorder = httptest.NewRecorder()
-	req := httptest.NewRequest("POST", "/api/v2/frontend/cost/calculate/file", bodyBuf)
-	req.Header.Add("Authorization", authHeader)
-	req.Header.Add("Content-Type", bodyWriter.FormDataContentType())
-	urlValues := url.Values{}
-	urlValues.Add("hold_time", "5")
-	req.PostForm = urlValues
-	api.r.ServeHTTP(testRecorder, req)
-	if testRecorder.Code != 200 {
-		t.Fatal("bad http status code recovered from /api/v2/frontend/cost/calculate/file")
-	}
-	floatAPIResp = floatAPIResponse{}
-	// unmarshal the response
-	bodyBytes, err := ioutil.ReadAll(testRecorder.Result().Body)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err = json.Unmarshal(bodyBytes, &floatAPIResp); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if floatAPIResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/frontend/cost/calculate/file")
-	}
-	if floatAPIResp.Response == 0 {
-		t.Fatal("failed to calculate cost /api/v2/frontend/cost/calculate/file")
-	}
-}
-
-func Test_API_Routes_Database(t *testing.T) {
-	// load configuration
-	cfg, err := config.LoadConfig("../../testenv/config.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	db, err := loadDatabase(cfg)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// setup fake mock clients
-	fakeLens := &mocks.FakeIndexerAPIClient{}
-	fakeOrch := &mocks.FakeServiceClient{}
-	fakeSigner := &mocks.FakeSignerClient{}
-
-	api, _, err := setupAPI(fakeLens, fakeOrch, fakeSigner, cfg, db)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// test database global uploads
-	// /api/v2/database/uploads
-	var interfaceAPIResp interfaceAPIResponse
-	if err := sendRequest(
-		api, "GET", "/api/v2/database/uploads", 200, nil, nil, &interfaceAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if interfaceAPIResp.Code != 200 {
-		t.Fatal("bad api status code from api/v2/database/uploads")
-	}
-
-	// test database specific uploads
-	// /api/v2/database/uploads/testuser
-	interfaceAPIResp = interfaceAPIResponse{}
-	if err := sendRequest(
-		api, "GET", "/api/v2/database/uploads/testuser", 200, nil, nil, &interfaceAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if interfaceAPIResp.Code != 200 {
-		t.Fatal("bad api status code from api/v2/database/uploads")
-	}
-
-	// test get encrypted uploads
-	// /api/v2/frontend/uploads/encrypted
-	if err := sendRequest(
-		api, "GET", "/api/v2/database/encrypted/uploads", 200, nil, nil, nil,
-	); err != nil {
-		t.Fatal(err)
-	}
-}
-
-func Test_API_Routes_Account(t *testing.T) {
-	// load configuration
-	cfg, err := config.LoadConfig("../../testenv/config.json")
-	if err != nil {
-		t.Fatal(err)
-	}
-	db, err := loadDatabase(cfg)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// setup fake mock clients
-	fakeLens := &mocks.FakeIndexerAPIClient{}
-	fakeOrch := &mocks.FakeServiceClient{}
-	fakeSigner := &mocks.FakeSignerClient{}
-
-	api, _, err := setupAPI(fakeLens, fakeOrch, fakeSigner, cfg, db)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// verify the username from the token
-	// /api/v2/account/token/username
-	var apiResp apiResponse
-	if err := sendRequest(
-		api, "GET", "/api/v2/account/token/username", 200, nil, nil, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if apiResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/account/token/username")
-	}
-	if apiResp.Response != "testuser" {
-		t.Fatal("bad username recovered from token")
-	}
-
-	// get an email verification token
-	// /api/v2/account/email/token/get
-	apiResp = apiResponse{}
-	if err := sendRequest(
-		api, "GET", "/api/v2/account/email/token/get", 200, nil, nil, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if apiResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/account/email/token/get")
-	}
-	// since we dont have email lets get the token from the database
-	um := models.NewUserManager(db)
-	user, err := um.FindByUserName("testuser")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if user.EmailVerificationToken == "" {
-		t.Fatal("failed to set email verification token")
-	}
-
-	// verify the email verification token
-	// /api/v2/account/email/token/verify
-	urlValues := url.Values{}
-	urlValues.Add("token", user.EmailVerificationToken)
-	apiResp = apiResponse{}
-	if err := sendRequest(
-		api, "POST", "/api/v2/account/email/token/verify", 200, nil, urlValues, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if apiResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/account/email/token/verify")
-	}
-	user, err = um.FindByUserName("testuser")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if !user.EmailEnabled {
-		t.Fatal("failed to enable email address")
-	}
-
-	// verify account password change
-	// /api/v2/account/password/change
-	urlValues = url.Values{}
-	urlValues.Add("old_password", "admin")
-	urlValues.Add("new_password", "admin1234@")
-	apiResp = apiResponse{}
-	if err := sendRequest(
-		api, "POST", "/api/v2/account/password/change", 200, nil, urlValues, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if apiResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/account/password/change")
-	}
-
-	// create ipfs keys
-	// /api/v2/account/key/ipfs/new
-	urlValues = url.Values{}
-	urlValues.Add("key_type", "ed25519")
-	urlValues.Add("key_bits", "256")
-	urlValues.Add("key_name", "key1")
-	apiResp = apiResponse{}
-	if err := sendRequest(
-		api, "POST", "/api/v2/account/key/ipfs/new", 200, nil, urlValues, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if apiResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/account/key/ipfs/new")
-	}
-	// test rsa keys
-	urlValues.Add("key_type", "rsa")
-	urlValues.Add("key_bits", "2048")
-	urlValues.Add("key_name", "key2")
-	apiResp = apiResponse{}
-	if err := sendRequest(
-		api, "POST", "/api/v2/account/key/ipfs/new", 200, nil, urlValues, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if apiResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/account/key/ipfs/new")
-	}
-	if apiResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/account/key/ipfs/new")
-	}
-	// manually create the keys since we arent using queues
-	if err = um.AddIPFSKeyForUser("testuser", "key1", "muchwow"); err != nil {
-		t.Fatal(err)
-	}
-	if err = um.AddIPFSKeyForUser("testuser", "key2", "suchkey"); err != nil {
-		t.Fatal(err)
-	}
-
-	// get ipfs keys
-	// /api/v2/account/key/ipfs/get
-	var mapAPIResp mapAPIResponse
-	if err := sendRequest(
-		api, "GET", "/api/v2/account/key/ipfs/get", 200, nil, nil, &mapAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if mapAPIResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/account/key/ipfs/get")
-	}
-
-	// get available credits
-	// /api/v2/account/credits/available
-	var floatAPIResp floatAPIResponse
-	if err := sendRequest(
-		api, "GET", "/api/v2/account/credits/available", 200, nil, nil, &floatAPIResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if floatAPIResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/account/credits/available")
-	}
-
-	// forgot email
-	// /api/v2/account/email/forgot
-	apiResp = apiResponse{}
-	if err := sendRequest(
-		api, "POST", "/api/v2/account/email/forgot", 200, nil, nil, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if apiResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/account/email/forgot")
-	}
-
-	// test@email.com
-	// forgot username
-	// /api/v2/forgot/username
-	apiResp = apiResponse{}
-	urlValues = url.Values{}
-	urlValues.Add("email_address", "test@email.com")
-	if err := sendRequest(
-		api, "POST", "/api/v2/forgot/username", 200, nil, urlValues, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if apiResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/forgot/username")
-	}
-
-	// test@email.com
-	// forgot password
-	// /api/v2/forgot/password
-	apiResp = apiResponse{}
-	urlValues = url.Values{}
-	urlValues.Add("email_address", "test@email.com")
-	if err := sendRequest(
-		api, "POST", "/api/v2/forgot/password", 200, nil, urlValues, &apiResp,
-	); err != nil {
-		t.Fatal(err)
-	}
-	// validate the response code
-	if apiResp.Code != 200 {
-		t.Fatal("bad api status code from /api/v2/forgot/password")
-	}
-}
 func Test_Utils(t *testing.T) {
 	cfg, err := config.LoadConfig("../../testenv/config.json")
 	if err != nil {
