+++ api/v2/routes_database.go
@@ -7,44 +7,15 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
-// GetUploadsFromDatabase is used to read a list of uploads from our database
-func (api *API) getUploadsFromDatabase(c *gin.Context) {
-	username, err := GetAuthenticatedUserFromContext(c)
-	if err != nil {
-		api.LogError(err, eh.NoAPITokenError)(c, http.StatusBadRequest)
-		return
-	}
-	if err := api.validateAdminRequest(username); err != nil {
-		FailNotAuthorized(c, eh.UnAuthorizedAdminAccess)
-		return
-	}
-	// fetch the uploads
-	uploads, err := api.upm.GetUploads()
-	if err != nil {
-		api.LogError(err, eh.UploadSearchError)(c, http.StatusInternalServerError)
-		return
-	}
-	api.l.Info("all uploads from database requested")
-	Respond(c, http.StatusOK, gin.H{"response": uploads})
-}
-
-// GetUploadsForUser is used to read a list of uploads from a particular user name
-// If not called by admin  admin, will retrieve all uploads for the current authenticated user
+// GetUploadsForUser is used to retrieve all uploads for the authenticated user
 func (api *API) getUploadsForUser(c *gin.Context) {
 	username, err := GetAuthenticatedUserFromContext(c)
 	if err != nil {
 		api.LogError(err, eh.NoAPITokenError)(c, http.StatusBadRequest)
 		return
 	}
-
-	var queryUser string
-	if err = api.validateAdminRequest(username); err == nil {
-		queryUser = c.Param("user")
-	} else {
-		queryUser = username
-	}
 	// fetch all uploads for that address
-	uploads, err := api.upm.GetUploadsForUser(queryUser)
+	uploads, err := api.upm.GetUploadsForUser(username)
 	if err != nil {
 		api.LogError(err, eh.UploadSearchError)(c, http.StatusInternalServerError)
 		return
